move support_data and configs into nested dir
