+++ config_pearl/configs/telescope.toml
