+++ config_pearl/support_data/coatings/AlMgF2_curve.csv
