+++ config_pearl/support_data/optical_designs/telescope+esc/STP_Mark_12F11_Coronagraph_HC11_1.zar
