+++ config_pearl/support_data/optical_designs/telescope+esc/readme.md
