+++ config_pearl/support_data/optical_designs/telescope+ifs/throughput_baseline.ecsv
