+++ config_pearl/support_data/optical_designs/telescope-only/INST-0200.pdf
