+++ config_pearl/support_data/optical_designs/telescope-only/STP_TMA_13_03.zprj
