+++ config_pearl/support_data/optical_designs/telescope-only/readme.md
